+++ lib/widgets/bottom_navbar.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:notion_wordbook/screens/add_wordbook_page.dart';
 import 'package:notion_wordbook/screens/home_page.dart';
 import 'package:notion_wordbook/screens/hook_test_page.dart';
-import 'package:notion_wordbook/screens/my_info_page.dart';
+import 'package:notion_wordbook/screens/test_result/test_result_page.dart';
 import 'package:notion_wordbook/screens/review_page.dart';
 import 'package:notion_wordbook/screens/user_page.dart';
 
@@ -23,7 +23,7 @@ class _State extends State<NavigatedPages> {
   final screens = [
     const HomePage(),
     const ReviewPage(),
-    const MyInfoPage(),
+    const TestResultPage(),
     AddWordbookPage(),
     const HookPage(),
     const MyPageScreen(),
